♻️ refactor:  dropdown error수정
